+++ cmd/eks-node-monitoring-agent/restconfig_test.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"os"
 	"path/filepath"
 	"testing"
 
@@ -43,23 +42,7 @@ func TestRewriteExecProvider(t *testing.T) {
 }
 
 func TestResolveCACertPath(t *testing.T) {
-	t.Run("embedded CA data needs no override", func(t *testing.T) {
-		hostRoot := t.TempDir()
-		t.Setenv(config.HOST_ROOT_ENV, hostRoot)
-		kubeconfigPath := writeKubeconfig(t, hostRoot, &clientcmdapi.Cluster{
-			CertificateAuthorityData: []byte("---embedded ca---"),
-		})
-
-		caCertPath, err := (&podRestConfigProvider{}).resolveCACertPath(kubeconfigPath)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if caCertPath != "" {
-			t.Fatalf("expected empty path for embedded CA data, got %q", caCertPath)
-		}
-	})
-
-	t.Run("referenced CA file is prefixed with the host root", func(t *testing.T) {
+	t.Run("resolves the CA of the kubeconfig's current context", func(t *testing.T) {
 		hostRoot := t.TempDir()
 		t.Setenv(config.HOST_ROOT_ENV, hostRoot)
 		kubeconfigPath := writeKubeconfig(t, hostRoot, &clientcmdapi.Cluster{
@@ -75,96 +58,9 @@ func TestResolveCACertPath(t *testing.T) {
 		}
 	})
 
-	t.Run("referenced CA file already inside the host root is left as-is", func(t *testing.T) {
-		hostRoot := t.TempDir()
-		t.Setenv(config.HOST_ROOT_ENV, hostRoot)
-		caCertPath := filepath.Join(hostRoot, "/etc/kubernetes/pki/ca.crt")
-		kubeconfigPath := writeKubeconfig(t, hostRoot, &clientcmdapi.Cluster{
-			CertificateAuthority: caCertPath,
-		})
-
-		got, err := (&podRestConfigProvider{}).resolveCACertPath(kubeconfigPath)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if got != caCertPath {
-			t.Fatalf("expected %q unchanged, got %q", caCertPath, got)
-		}
-	})
-
-	t.Run("falls back to the well-known host location", func(t *testing.T) {
-		hostRoot := t.TempDir()
-		t.Setenv(config.HOST_ROOT_ENV, hostRoot)
-		wellKnown := filepath.Join(hostRoot, "/etc/kubernetes/pki/ca.crt")
-		if err := os.MkdirAll(filepath.Dir(wellKnown), 0o755); err != nil {
-			t.Fatal(err)
-		}
-		if err := os.WriteFile(wellKnown, []byte("---ca---"), 0o644); err != nil {
-			t.Fatal(err)
-		}
-		kubeconfigPath := writeKubeconfig(t, hostRoot, &clientcmdapi.Cluster{})
-
-		got, err := (&podRestConfigProvider{}).resolveCACertPath(kubeconfigPath)
-		if err != nil {
-			t.Fatalf("unexpected error: %v", err)
-		}
-		if got != wellKnown {
-			t.Fatalf("expected %q, got %q", wellKnown, got)
-		}
-	})
-
-	t.Run("errors when no CA can be located", func(t *testing.T) {
-		hostRoot := t.TempDir()
-		t.Setenv(config.HOST_ROOT_ENV, hostRoot)
-		kubeconfigPath := writeKubeconfig(t, hostRoot, &clientcmdapi.Cluster{})
-
-		if _, err := (&podRestConfigProvider{}).resolveCACertPath(kubeconfigPath); err == nil {
-			t.Fatal("expected an error when no CA can be located")
-		}
-	})
-
 	t.Run("errors when the kubeconfig cannot be loaded", func(t *testing.T) {
 		if _, err := (&podRestConfigProvider{}).resolveCACertPath("/does/not/exist"); err == nil {
 			t.Fatal("expected an error for a missing kubeconfig")
-		}
-	})
-}
-
-func TestClusterForCurrentContext(t *testing.T) {
-	t.Run("resolves the cluster referenced by the current context", func(t *testing.T) {
-		want := &clientcmdapi.Cluster{Server: "https://current"}
-		kubeconfig := &clientcmdapi.Config{
-			CurrentContext: "ctx",
-			Contexts:       map[string]*clientcmdapi.Context{"ctx": {Cluster: "wanted"}},
-			Clusters: map[string]*clientcmdapi.Cluster{
-				"wanted": want,
-				"other":  {Server: "https://other"},
-			},
-		}
-		if got := clusterForCurrentContext(kubeconfig); got != want {
-			t.Fatalf("expected %+v, got %+v", want, got)
-		}
-	})
-
-	t.Run("falls back to the sole cluster when the context is missing", func(t *testing.T) {
-		want := &clientcmdapi.Cluster{Server: "https://only"}
-		kubeconfig := &clientcmdapi.Config{
-			Clusters: map[string]*clientcmdapi.Cluster{"only": want},
-		}
-		if got := clusterForCurrentContext(kubeconfig); got != want {
-			t.Fatalf("expected the sole cluster, got %+v", got)
-		}
-	})
-
-	t.Run("returns nil when the context is missing and clusters are ambiguous", func(t *testing.T) {
-		kubeconfig := &clientcmdapi.Config{
-			Clusters: map[string]*clientcmdapi.Cluster{
-				"a": {Server: "https://a"},
-				"b": {Server: "https://b"},
-			},
-		}
-		if got := clusterForCurrentContext(kubeconfig); got != nil {
-			t.Fatalf("expected nil for ambiguous clusters, got %+v", got)
 		}
 	})
 }
